Moved fileset upgrade to Upgrader Tool
